Update business object conversions
